Monta t 32 (#42)
* Update .gitignore

* Create docker-compose.yml

* Update gitignore and requirements

* Fix mypy issues

* Add new Views to Workers APP

* Update models.py

* Small changes

* Update api.py

Query optimization for accounting api.

* Remove DB constraint and change to clean.

Removal of unique DB constraint on account number. It was firing before the update method was being called on the serializer, so in return created validation in the clean method to ensure that the account number is unique, but will allow for the user to update the record.

* Changes to Account API

* Code Cleanup

* Change Utils

* Update Worker Serializers

* Location Field Change

Make location field nullable on stop model. If company does not have the location added to the system and it's a one off, give the ability to just add the address line itself.

* Mass Update to serializers

Changed the serializers to PrimaryKeyRelatedFields. Want to give the users the ability to add existing records to the system instead of having to create new ones. If they want to add new codes, there can be a modal or different page to add master data.

* Update urls.py

Add customer billing profile.

* Start of Orders API.

* Order Management

Finalize the orders api with movements, stop and additional viewsets required to complete an order.

* Add filters to Orders API

Adding some filterset_fields on the orders api viewssets. This will make it easier for the frontend end developer to do filtering on specific queries.

* add filters to stop and movement api.

* Model fixes.

* [pre-commit.ci] auto fixes from pre-commit.com hooks

for more information, see https://pre-commit.ci

* Finished API Endpoints, changed organizationViewSet to OrganizationMixin

* [pre-commit.ci] auto fixes from pre-commit.com hooks

for more information, see https://pre-commit.ci

* Update serializers.py

* Update mypy errors

* Mypy and tests fixes

* [pre-commit.ci] auto fixes from pre-commit.com hooks

for more information, see https://pre-commit.ci

Co-authored-by: Eric Moss <yeswolfprod@gmail.com>
Co-authored-by: pre-commit-ci[bot] <66853113+pre-commit-ci[bot]@users.noreply.github.com>