Movement Tests
Will have to come back to this tests. If you create a movement out right without stops, the API fails. That is primarily due to adding an order does the initial work ,but if a user wants to add a one off-stop they will need to add it to the movement the is created on order generation. The user needs to be able to add stops to other movements.