Movement tests (#44)
* fix imports due to file move

* Movement Tests

Will have to come back to this tests. If you create a movement out right without stops, the API fails. That is primarily due to adding an order does the initial work ,but if a user wants to add a one off-stop they will need to add it to the movement the is created on order generation. The user needs to be able to add stops to other movements.

* Mypy correction

* [pre-commit.ci] auto fixes from pre-commit.com hooks

for more information, see https://pre-commit.ci

Co-authored-by: Eric Moss <yeswolfprod@gmail.com>
Co-authored-by: pre-commit-ci[bot] <66853113+pre-commit-ci[bot]@users.noreply.github.com>