CHANGE: Order Model
Moved Validation back into the class Order clean method. Removed auto_pop_address from OrganizationControl. It made sense for this to be set by default. There shouldn't be a reason a customer would not want this. There is a contingency to prevent the order address from auto populating from the location ID if the customer puts information into the address_line directly otherwise the Location information will be auto populated.