CHANGE: Order Admin
Changed Order Admin to Nested Admin Inlines. This change will help visually to ensure that you are able to see what is tied to what rather than viewing separate admin links.

Add Celery Tasks to consolidate PDF. This tasks is heavy depending on how many documents need to be combined into one. Made sense for this to be ran in celery rather than a signal or in a view.

Add total_order_weight and total_order_piece_count to order signals. This computation used to be done directly in the model. However after running some tests on this. It made more sense for this to be done in a signal rather than in the model save method. May take another look at this in the future to see if it needs to be moved into a celery tasks.

Change ordering on Order Documentation and Order Comment Model.