CHANGE: Stops Selector & Pytest.raises.
Changed total piece & weight to selectors instead of being a special method in the Order Model.

Changed Pytest.raises(match=etc...), I was informed by one of the maintainers of the project that this throws false positives and after doing some further testing he was correct. So I'm taking execution information and doing assertions on the message directly within the JSON object to ensure that the Validation is being thrown ,but also the message is a direct match.