CHANGE: Fix Accounts Tests & Serializer
Found an issue with the Accounts Serializer not passing organization in the profile_data. Due to this field being ignored on the GenericSerializer class. You must add the field to profile_data which grabs the organization from the requested user.

Updated the Accounts tests with these adjustments, and changed the pytest.raises on the admin user creation validation tests.