CHANGE: Remove Fields from Order Control & Update Orders Model
Change help_text on enforce_customer field, generate_routes on OrderControl Model

Removed enforce_shipper on OrderControl Model

Updated Clean() method on Order model.

Changed comments to be more descriptive.

Add validation to enforce revenue_code if the OrderControl for the organization has the enforce_revenue_code field as true.

Changed auto_order_total to the clean method and made changes to calculate_total to only be called if the OrderControl has this turned on.