FIX: Movement Entry creating multiple stops
Fixed bug where even if the movement was not the initial movement(Movement created from entering origin and destination in the Order) the additionally movements entered would trigger the create_initial stops signal to be fired adding additional stops to the movement. To fix this, added a query to the signal to get the query of all of the movements in the order, if the order is created and the movement count for the order is equal to 1, then the create_initial_stops signal will be fired. Otherwise it will be ignored.