MONTA-T-34 (#54)
* CHANGE: Token API

Move logic to create a token to the serializer instead of the post method on the
View.

- Add password to the serializer as WRITE_ONLY

- Add extend_schema_serializer decorator to ensure that the generated documentation gives the right response example.

* FIX: Enum Overrides

Fix error with drf-spectacular throwing warning for LicenseStateEnum from localflavor

* UPDATE: Urls

Fix typo in change-password route.

* UPDATE: Allowed HTTP Methods for OrderControl

Since the Organization can only have one OrderControl, and this is automatically generated on the Creation of an organization, I have removed the ability to send a post request to this endpoint.

* FIX: Organization Warning

FIX: Issue with organization_pk throwing warning from drf-spectacular.

* CHANGE: Token & Accounts API & Test

- Change Token API to validate from the serializers ,but create the operations in the View.
- Change the param 'user_name' to 'username' in the create_user function.
- Add validation & post tests to Token API.
- Re-organization default_renderer for DRF
- Add tests for posting user.

* CHANGE: Add exclusion for tests files in mypy.

* FIX: Delete operations now return status_code 200.

Fix tests to assert the proper status code for the delete operation.

* CHANGE: Accounts serializer & Tests

- Add example to proper example to User api endpoint.
- Add department__name to filterset instead of department ID.

* CHANGE: Remove unused imports

* [pre-commit.ci] auto fixes from pre-commit.com hooks

for more information, see https://pre-commit.ci

Co-authored-by: Eric Moss <yeswolfprod@gmail.com>
Co-authored-by: pre-commit-ci[bot] <66853113+pre-commit-ci[bot]@users.noreply.github.com>