CHANGE: BillingQueue save
Removed the bulk_create from the BillingQueue creation when transferring an order. The bulk_create operation does not call the model save method, so to mitigate this, I decided to do a for loop for each record and create them from there. The bulk create does provide better performance ,but this tasks will be ran from celery, so I don't believe it matters at this point.