WIP: syntax parser for Django models
This code may never make it to prod ,but I'm going to leave it here for now. All it does is take some custom syntax and creates the syntax for a django model. Idea is to take the syntax in, parse it, and add it to an django applicaitons models.py file. Cool idea, but may be a risk to it. Will have to think about syntax errors, etc....