(WIP) Plugin Application
Working on adding a feature to Monta that allows customers to installs certain "plugins" to their instance. This will be used in the future to provide customers with additional application to Monta, instead of having to go into each instance and add the applications by hand. The system administrator can simply be given a temporary link to a zip file from Github, or the account manager can go into the account and install the application to the instance.