add Header menu close
Unfortunately, Mantine does not provide a way to control the opened or close state on a hovercard. So instead we simply increment a key which then force react to identify an item has changed, been added or removed. This way when a link is clicked, the menu will disappear, not the best solution ,but it works and doesn't seem to  have any performance issues.