add: returnURL to login page
return url is used to return the user back to the page they were on before they logged out, so if they were to logout and back in, it will take them back to the page they were on previously. Once the user successfully logs in, the returnURL session item will be removed.