chore: update to react-query v5
make changes to queries and put all redundant queries in one file instead of all in separate files. Additionally, you'll notice that each query key has `as QueryKeys[]` this is on purpose. The purpose of this is QueryKeys is a type that contains all of the available query keys for the application. If they ever change I want to ensure an error is raised so that it can be fixed. Nothing like pulling or invalidating from a query that does not exist.