fix: layout errors
Due to the migration to React Query v5 that caused the majority of the useQuery hooks to throw type errors, however since the user information is grabbed in many places a hook of our own was created to get the information. Additionally, some small style were made to the layout skeleton and navigation menu.