add: custom default DRF pagination
This was something that should have been added a long time ago. On the client application I'm having to set the limit to a large number for a lot of our select-input that use server-side data due to a query param for all records not be available this should resolve that issue.