add: test and validation for trailer availability status
Added a new validation rule and corresponding test to check the availability status of a trailer before a movement is created. Now, a ValidationError is thrown if the status of the trailer is not 'available'. This is meant to prevent the assignment of a trailer to a movement if it isn't available. Updating equipment models and removing a redundant print statement were also parts of this commit