change: simplify code and remove 'description' from models and admin interface
This commit refactors the admin.py and models.py files in the server/equipment directory. It simplifies the declaration of model attributes and the list_display and search_fields tuples in the admin interface classes. It also removes the 'description' attribute from the Tractor, Trailer and EquipmentMaintenancePlan models and their corresponding admin interface classes. The help_text of various fields in the Tractor model has also been updated for clarity.