add: shipment map view (#192)
* fix: font issue

* change: shipment store

Instead of just passing the ID of the worker into the Shipment store we send the entire json object, so that we have the ability to retrieve and use any field that we want from it.

* wip: view hos logs

* change: notification menu to load lazy image

* Update app-grid.tsx

* add: keybinds to user avatar menu

* change: shipment map view

* Update Shipment.tsx

* Update ShipmentStore.ts

* add: tractor to exportable models

* Update shipment-breadcrumb.tsx

* change: persist map preferences in local storage

* add(organization): feature flag

added feature flag at the organization level. Doesn't really make much sense to apply this at the business unit level due to the fact that Monta is scoped out to support brokerages and possibly other forms of transportation that may not have the same features applied to them.

* Update api.py

* wip: organization feature flags

* add: timezone to user model

* chore: cleanup deps

* Update App.tsx

* fix: not clearing input on close

When closing the site search with an invalid route it would leave the previously search text, but wouldn't display the error letting the user know the route didn't exist. Fixed by clearing the input when the dialog is closed

* add: list of current feature flags for Monta

* add: accessibility

* [pre-commit.ci] auto fixes from pre-commit.com hooks

for more information, see https://pre-commit.ci

* chore: bump

* [pre-commit.ci] auto fixes from pre-commit.com hooks

for more information, see https://pre-commit.ci

* remove waffle middleware

---------

Co-authored-by: pre-commit-ci[bot] <66853113+pre-commit-ci[bot]@users.noreply.github.com>