change: add more exclusion to get table names
We don't want to give access to every table for users to be able to pull from. This may change later based on requirements from customers. They can vote on which tables to add or remove. Some of them will never be able to be added due to security reasons.