feat: control file admin page (#200)
* change: checkbox to be clickable anywhere

Instead of having to click directly onto the checkbox, you can click on the label, description, etc... and it will check the checkbox.

* wip: control files

* feat: shipment control page

* Update README.md

* add: route control page

* add: feasibility control

* change: add more exclusion to get table names

We don't want to give access to every table for users to be able to pull from. This may change later based on requirements from customers. They can vote on which tables to add or remove. Some of them will never be able to be added due to security reasons.

* chore: cleanup types

* fix: control file responsiveness

* Update test_dispatch_control.py

* Update factories.py