add: app name change (#207)
* add: conditional logic validation

* change: conditional logic type dict

* add: conditional logic example

* remove print from conditional logic model

* add: validation for model field

when adding conditional logic ensure the model field the user is using is apart of the model in some form.

* add: test for validating model fields

* wip: conditional logic

so far this works some what, definitely need to do more testing.

* wip: table change conditionals

* wip: table change alert conditional logic

* fix: wrong column name

* fix: minor changes to organization apis

* Update test_table_changes.py

* change: add operation enum

* add: get routine definition

this is helper that can be used to get the routine definition from the database. this is not be used in anything that is external facing, and should only be used for tests.

* add: test for routine definition

ensure the routine definition contains the proper information

* chore: bump deps

* remove: strict mode

unfortunately, this has to go. It was breaking stuff that I can't fix because they are third party packages like the markers in the shipment map view.

* chore: bump deps

* fix: shipment map view

Minor adjustment as this product is moving to staging, to remove the dependency on having the API key on the client, rather using the organization api key to get the information.

* add: google maps admin page

* add: alert to google api page

* fix: login page text

* init: app name change

* backend: app name change

* init: client rename

* update: copyright

* final: application rename

* final

* [pre-commit.ci] auto fixes from pre-commit.com hooks

for more information, see https://pre-commit.ci

* minor changes

* [pre-commit.ci] auto fixes from pre-commit.com hooks

for more information, see https://pre-commit.ci

* Update test_location.py

* fix: test_generate_thumbnail test

* [pre-commit.ci] auto fixes from pre-commit.com hooks

for more information, see https://pre-commit.ci

---------

Co-authored-by: pre-commit-ci[bot] <66853113+pre-commit-ci[bot]@users.noreply.github.com>