✅(change): enable comment for entities
We'll be querying the comment for each individual column within a table to provide a better overall experience to the user on the client. The BaseMixin will allow comments, but individual fields within the entities will need comments added to them. Before this is moved into production, we'll need to use redis to cache the response from the "columns" api to ensure that we're not hitting the database on every request.