✨(feature): customized reporting module (#249)
* ✅(add): get columns by table name api

* 🍪(chore): code cleanup

* ✅(change): improve client impls

* Update table_change_alert.service.go

* ✨(add): column names api

* 🍪(chore): formatting improvement

* Update server_config.go

* 🔴(fix): force atlas.sum

* ✅(chore): replace atlas ci with cloud

* Update ci-atlas.yaml

* Delete atlas.sum

* Delete 20240411001537_add_unique_indexes.sql

* Update ci-atlas.yaml

* 🔴(fix): cache on lint workflow

* Update router.go

* ✅(add): user notification entity

* Update ci-atlas.yaml

* Update test.yml

* Update test.yml

* 🪛(fix): docker compose file

* 🟡(wip): reporting service

I've added the reporting micro-service in a separate repository. It is working as intended and return back the report url. However we still need to generate the notification entity.

* ✅(change): enable comment for entities

We'll be querying the comment for each individual column within a table to provide a better overall experience to the user on the client. The BaseMixin will allow comments, but individual fields within the entities will need comments added to them. Before this is moved into production, we'll need to use redis to cache the response from the "columns" api to ensure that we're not hitting the database on every request.

* 🟡(wip): reporting service

* 🪛(fix): ci

* 🍪(chore): bump deps

* ✨(feat): report service