✨(feature): google maps integration (#252)
* 🟡(change): replace react-hot-toast with sonner

* Update App.tsx

* 🟡(change): authentication page

Authentication will now check if the user apart of an organization, and then proceed to allow this user to input their password.

* 🟡(change): light mode badge styling

* ✨(add): haversine algorithm

Big credit to https://github.com/umahmood/haversine/ . This of course was re-implemented from our old python application to meet the needs of this application. This will provide a free alternative to the Google distance matrix for organization.

* ✨(add): google api client

* ✨(add): shipment route entity

* 🟡(change): replace haversine with Vincenty formula

* 🟡(change): improve vincenty distance

* 🪛(change): yeah I'm dumb

and yes this was revoked

* 🟡(change): move queries into model folder

* 🟡(change): improve kafka client impl

* 🍪(chore): improve documentation

* Update auth.handler.go

* 🍪(chore): bump