✨(feat): customer application (#254)
* 🪛(fix): form resets not occurring

Previous PR caused the forms to not reset once the data was submitted successfully.

* bump

* ✨(feat): report service relationship support

Adding the ability to get information from tables that are directly related to the table that is being queried. For example, if the user wishes to pull a report for revenue_codes, they will also be able to get the expense & revenue account information associated.

* 🍪(chore): bump

* 🟡(change): add indexes to delivery slot

* 🟡(change): font to geist

* 🪛(fix): make file ent path

* 🟡(change): user notification to only return error

there was no need to return back the notification entity so instead just return an error if there is one.

* 🍪(chore): seeder data

* 🟡(change): add icon & fix group name

* 🪛(fix): user perm. context causing re-renders

* 🍪(chore): cleanup

* 🪛(fix): table change alert dep array

* 🪛(fix): navbar re-renders

* 🍪(chore): cleanup

* 🟡(change): add organization id to update services

We have to pass the organization id so that we can validate the organization id vs a unique value. Even though the database may have constraints it is still nice to be able to do this in the app level

* 🍪(chore): go generate

* ✅(add): validate uniqueness helper

helper function that allows validate on the server by calling the database for specific columns

* 🪛(fix): seeder error

* ✨(feat): customer application