✨(feat): Table Change Alerts (#260)
* ✅(add): conditional logic validation

* 🟡(change): type definitions

* 🪛(fix): structure

* ✅(add): table change alert for database insert

* 🟡(change): key & value converter

Change the key & value converters for Kafka to JSON serializer. Unfortunately, even if we register the schema's in something like `goavro` it isn't able to deserialize the data due to the debezium types not being registered properly, rather than doing conversions on each type. Debezium recommends removing serialization all together or using the JSON serailizer if the information is needed. In this case, we need the values from the connector to be able to utilize table change alerts properly.

* 🍪(chore): cleanup

* 🍪(chore): improve logging