🟡(change): organization handler to bypass redis
we do not want to return err in this particular part of the code. The bug was that when I would refresh the application and flush the database the old organization ID would be fetched rather than the new one from redis. This shouldn't happen in production and seems to only be an issue on local redis instances, which the `FLUSHALL` command in redis seems to rectify.