🟡(change): audit service (#277)
* ✅(add): encrypt cookie middleware

* bump

* ✅(add): test for commodity codes

* bump

* bump

* bump

* bump

* bump

* 📍(wip): shipment control w/ validation

* Delete commodity.service_test.go

* 🍪(chore): fix seeder

* 🪛(fix): fix incorrect constants

* 🪛(fix): form not reloading properly

* When saving the organization and shipment control forms the form would not reset with the new data.
* When changing the organization logo and then saving the logo would go away and be overwritten with the previous logo and/or blank logo if one did not exist.

* ✨(feat): add shipment control

* 🟡(change): onSettled param to include response data

* 🍪(chore): cleanup deadcode

* 🟡(change): organization handler to bypass redis

we do not want to return err in this particular part of the code. The bug was that when I would refresh the application and flush the database the old organization ID would be fetched rather than the new one from redis. This shouldn't happen in production and seems to only be an issue on local redis instances, which the `FLUSHALL` command in redis seems to rectify.

* 🍪(chore): bump

* bump

* bump

no explaination. it is what it is.

* ✨(feat): new audit service