✅(add): query invalidation hook
Due to trenova utilizing popouts it is important that we send broadcast messages to invalidate the application state. for example if a user has a new tab open and they create a new record in the table, that record will not appear in a different tab, so we need to broadcast that message so all tabs know to invalidate and refetch the query results.