V523 (#324)
* 🍪(chore): cleanup

* 🪛(fix): move assignment not working properly

React hook form was overriding the default id of the movement which in return was sending the wrong id in the request to the backend. To fix this we simply added `formId` to the schema, so that react-hook-form has something to append the id to, even though we will not use this value on the backend.

This issue is being discussed here since 2022, and yet they haven't fixed it.

https://github.com/orgs/react-hook-form/discussions/8935

* 🪛(fix): assignment option disabled on wrong statuses

* 🍪(chore): bump

---------

Signed-off-by: Eric Moss <66630775+emoss08@users.noreply.github.com>