✅(add): rabbitmq
RabbitMQ will manage the communication between microservices and the main application. I started to notice that a lot of the things we'd normally bootrap to the main application we're causing issues, or would not normally be inside of application ,but a separate service entirely. For example, we've added `Hatchet` to the application which will facilitate long running processes in the application. Rather than bootstrapping this to the application, the main application will send messages using rabbitMQ to the service that manages these processes, as we do not want the main application to fail if something goes wrong in Hatchet.