✅(add): workflow microservice
The workflow microservice is responsible for managing jobs and facilitating long running processes within the application. Organizations can opt out of using the workflow service, by simply not starting it. The main application connects directly to RabbitMQ to facilitate communication to the workflow service, so messages will just be queued up, without the service running anyway.