🚀(release): v0.5.7 (#341)
* 🍪(chore): bump

* 🟡(change): compiler

* 🟡(change): document preview generator

* 🪛(fix): performance problems

* Update service.go

* ✅(add): river queue

* ✅(add): rabbitmq

RabbitMQ will manage the communication between microservices and the main application. I started to notice that a lot of the things we'd normally bootrap to the main application we're causing issues, or would not normally be inside of application ,but a separate service entirely. For example, we've added `Hatchet` to the application which will facilitate long running processes in the application. Rather than bootstrapping this to the application, the main application will send messages using rabbitMQ to the service that manages these processes, as we do not want the main application to fail if something goes wrong in Hatchet.

* ✅(add): gitignore to microserivce

* ✅(add): workflow microservice

The workflow microservice is responsible for managing jobs and facilitating long running processes within the application. Organizations can opt out of using the workflow service, by simply not starting it. The main application connects directly to RabbitMQ to facilitate communication to the workflow service, so messages will just be queued up, without the service running anyway.

* ✅(add): boilterplate code for workflow publiher

* 🪛(fix): error with log entry being marshalled

* Delete queue.go

* ✅(add): auto delay shipment workflow

* bump

* 🟡(change): show stop error in timeline

resolves #342

* 🟡(change): registration of workflow types.

* 🟡(change): inconsistencies

* 🍪(chore): remove unused gh action

* 🪛(fix): assignment not appearing

when a user would submit an assignment they would have to refresh the page on an existing shipment for the assignment to appear in the move timeline.

* 🪛(fix): stop times allowed without assignment

resolve #345

* 🍪(chore): bump

* 🟡(change): remove workflow service dependency

remove workflow service dep from shipment repository

* 🟡(change): add db ping to workflow service

* 🟡(change): add dead letter exchange

add dead latter exchange to prevent infinite re-queue on panic

* 🍪(chore): fix imports

* 🪛(fix): handle RabbitMQ queue precondition failures gracefully

* Update rabbitmq.go

* 🪛(fix): dlx

* 🪛(fix): shipment validation tests

* ✅(add): new validation framework

* 🍪(chore): bump deps

* 🟡(change): improve validation framework

improve the framework to account for nested error handling

* Create README.md

* 🟡(change): validators to use framework

* 🍪(chore): add golangci gh action

* Update test.yml

* Update test.yml

* 🟡(change): refactor validators

* Update tractor.go

* 🍪(chore): modernize

* 🍪(chore): bump

* Update base.go