🚀(release): v0.7.2 (#385)
* 🍪(chore); bump

* 🟡(change): add asynqmon to local compose

* ✅(add): convert nil pulid helper

* ✅(add): job service

* ✅(add): get all to shipment repo

* 🪛(fix): tractor repo test

* 🍪(chore): add no lint

* 🟡(change): primary worker id requirement

only require private worker id if auto assign is enabled.

* 🔴(remove): permission check from suggestion service

* 🟡(change): fix test utils

* ✅(add): dedicated lane suggestion API

* 🔴(remove): metrics middleware

* 🟡(change): improve pattern config

* 🍪(chore): improve comments

* 🍪(chore): format script files

* ✅(add): pattern config module

* 🟡(change): add count support to duplicate shipment

impl #378

* 🪛(fix): panic

fixes panic when no dedicated lane is found for a particular shipment

* 🟡(change): return err instead of nil

* 🟡(change): add handler to job module

* ⭕(wip): dedicated lane suggestions

* 🟡(change): duplicate shipment form

* 🟡(change): billing profile document types

impl #379

makes it easier to keep track of document types rather than just using a
json object

* ✅(add): system user fixture

system user will auto migrate with the application along with the admin
user. The system user will be used for background tasks, so that we can
log them.

* ✅(add): dedicated lane suggestion actions

* 🟡(change): vite config to rolldown-vite

* 🟡(change): improve session implementation

* 🟡(change): improve redis client impl

use JSON instead of string

* 🪛(fix): shipment test

* 🔴(remove): organization store

not being used

* 🔥(feat): add notification service

* ✅(add): add context to request

helper function to set tenant information to a struct

* 🟡(change): toaster style

* 🪛(fix): shipment not showing properly based on URL

* 🍪(chore): bump

* ⭕(wip): advanced table filtering

adding advanced table filtering to respective tables, such as customer,
shipment, trailer, tractors, workers and locations.

* 🍪(chore): bump claude code

* 🍪(chore): bump deps

* 🟡(change): override react-table types

* 🔥(feat): notifications

* ⭕(wip): shipment management v2

* 🟡(change): application style changes

* 🟡(change): add postgres search to query builder

deprecating the only postgressearch in favor of this. I will be leaving
the other function in the mean time while I build this out to my liking
utilizing the customer and shipment

* 🟡(change): filter options

* ⭕(wip): advanced table filtering

* 🔥(feat): add nested field support to query builder

ref #383

* 🍪(chore): bump

* 🟡(change): application logo