🔴(remove): resource editor page
we'll be rewriting this. The ace editor is heavy and isn't really need
for what we want to do. In the future we'll provide our own SQL editor
that is specific to our implementation. For now, users will need to
connect directly to their database to get information or make API calls
to the resource editor endpoint.